Support adding events with externalId or sessionId
When the customerId is not known, events can be created sending an
externalId or a sessionId instead. The API requires these kind of ids to
be wrapped in a "bringBackProperties" object. To provide a better
experience to the SDK user, I abstracted this object and provided two
top-level properties instead. The bringBackProperties object is created
on the fly during the serialization of the Event.